restructure and fix headers
