+++ src/annotation/annotation.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <optional>
 #include <string>
 #include <string_view>
 #include <unordered_map>
