+++ src/core/color.h
