--- src/internal/time/datetime.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <cstdint>
-#include <regex>
-#include <string>
-
-namespace ChessGame
-{
-    using std::chrono::utc_time;
-} // namespace ChessGame
